Moved send_email_ses() to boto_manager.py
